requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyyaml>=6.0
python-dateutil>=2.8.0
